+++ src/config/default.js
@@ -64,7 +64,7 @@ export default {
     latestSeasonCacheTtl: hours(1),
     refreshInterval: minutes(30),
     getCacheTtl: minutes(30),
-    backupLatestSeasonId: '14E58A5B-9139-4B9B-A0B7-D2B79E303332',
+    backupLatestSeasonId: 'B8823ECB-6E70-4D6F-BFC9-45768FA3C6B2',
   },
 
   cache: {